+++ docs/changes.md
@@ -1,5 +1,8 @@
 # 最新动态
 
+2023/11/19
+  * 完善group box，支持value属性，方便MVVM绑定radio button。
+
 2023/11/18
   * 修复STM32上编译问题。
   * remote\_ui\_service支持设置find\_target回调函数。